requirements.txt
APScheduler==3.11.2
bcrypt==5.0.0
blinker==1.9.0
boto3==1.42.70
botocore==1.42.70
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.7
Flask==3.1.3
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
greenlet==3.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.1.0
MarkupSafe==3.0.3
mysql-replication==1.0.15
packaging==26.1
pycparser==3.0
PyJWT==2.11.0
PyMySQL==1.1.2
pymysqlblinker==1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
s3transfer==0.16.0
six==1.17.0
SQLAlchemy==2.0.48
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
urllib3==2.6.3
Werkzeug==3.1.6
